requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
pymysql
cryptography
redis
pymongo
numpy
pycryptodome
werkzeug
python-dotenv
gunicorn
requests
jsonschema